+++ Aula 03/exercicio 08/cadastroCursos.js
@@ -67,20 +67,44 @@ function atualizarCurso(){
 function excluirCurso() {
 
     const id = readline.questionInt("Digite o ID do CURSO: ");
+    const sql = "SELECT * FROM cursos WHERE id = ?";
 
-    const deletar ="DELETE FROM cursos WHERE id = ?";
+    conexao.query(sql,[id], function(erro, cursos){
+        if(erro){
+            console.log("Erro ao buscar curso.");
+            console.log(erro)
+            menu();
+        } else if(cursos.length === 0){
+            console.log("Curso não encontrado. ");
+            menu();
+        } else{
+            console.log("\nCurso encontrado: ");
+            console.log("Nome: ", cursos[0].nome);
+            console.log("Carga horaria: ", cursos[0].carga_horaria);
 
-    conexao.query(deletar, [id], function (erro, resultado){
+            const confirmar = readline.question("Deseja realmente excluir este curso? (S/N) ");
 
-        if (erro) {
-            console.log("Erro ao excluir o curso.");
-        } else if (resultado.affectedRows === 0) {
-            console.log("Curso não encontrado.");
-        } else {
-            console.log("Curso excluído com sucesso!");
+            if(confirmar === "S" || confirmar === "s"){
+                const deletar = "DELETE FROM cursos WHERE id = ?";
+                conexao.query(deletar,[id],function(erro,resultado){
+                    if(erro){
+                        console.log("Erro ao excluir curso.");
+                        console.log(erro)
+                    } else if(resultado.affectedRows === 0){
+                        console.log("Curso não encontrado. ")
+                    } else {
+                        console.log("Curso excluido com sucesso! ")
+                    }
+                    menu();
+                });
+            }else if(confirmar === "N" || confirmar === "n"){
+                console.log("Exclusão cancelada.");
+                menu();
+            }else{
+                console.log("Opção inválida. ");
+                menu();
+            }
         }
-
-        menu();
     });
 }
